+++ lucabook/lib/luca_book/setup.rb
@@ -9,7 +9,7 @@ module LucaBook
     def self.create_project(country = nil, dir = LucaSupport::Config::Pjdir)
       FileUtils.mkdir_p(dir) unless Dir.exist?(dir)
       Dir.chdir(dir) do
-        %w[data/journals dict].each do |subdir|
+        %w[data/journals data/balance dict].each do |subdir|
           FileUtils.mkdir_p(subdir) unless Dir.exist?(subdir)
         end
         dict = if File.exist?("#{__dir__}/templates/dict-#{country}.tsv")
@@ -18,6 +18,21 @@ module LucaBook
                  'dict-en.tsv'
                end
         FileUtils.cp("#{__dir__}/templates/#{dict}", 'dict/base.tsv') unless File.exist?('dict/base.tsv')
+        prepare_starttsv(dict) unless File.exist? 'data/balance/start.tsv'
+      end
+    end
+
+    # Generate initial balance template.
+    # Codes are same as base dictionary.
+    # The previous month of start date is better for _date.
+    #
+    def self.prepare_starttsv(dict)
+      CSV.open('data/balance/start.tsv', 'w', col_sep: "\t", encoding: 'UTF-8') do |csv|
+        csv << ['code', 'label', 'balance']
+        csv << ['_date', '2020-1-1']
+        CSV.open("#{__dir__}/templates/#{dict}", 'r', col_sep: "\t", encoding: 'UTF-8').each do |row|
+          csv << row if /^[1-9]/.match(row[0])
+        end
       end
     end
   end