+++ lucabook/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Handle starting balance.
 * LucaBook::Import append 'x-editor: LucaBook::Import' header.
+* provide data/balance/start.tsv template on Setup.
 
 ## LucaBook 0.2.15